Switch over faucet command too
